Rename alternate bundler package name
Test Plan: CI. Tested locally with `test/development/app-dir/basic/basic.test.ts`. Will test with a create-next-app once published.
