+++ packages/next-rspack/README.md
@@ -1,4 +1,4 @@
-# @next/plugin-rspack (EXPERIMENTAL)
+# next-rspack (EXPERIMENTAL)
 
 > [!WARNING]
 > This package is currently experimental and actively developed and supported in Next.js’ `canary` branch. To use this, you must be using a published canary build of Next.js.
@@ -8,13 +8,13 @@ This plugin allows you to use [Rspack](https://rspack.dev) in place of webpack w
 ## Installation
 
 ```
-npm install @next/plugin-rspack
+npm install next-rspack
 ```
 
 or
 
 ```
-yarn add @next/plugin-rspack
+yarn add next-rspack
 ```
 
 ## Usage
@@ -22,7 +22,7 @@ yarn add @next/plugin-rspack
 Create or update a `next.config.js`/`next.config.ts` and wrap your existing configuration:
 
 ```js
-const withRspack = require('@next/plugin-rspack')
+const withRspack = require('next-rspack')
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
@@ -34,11 +34,11 @@ module.exports = withRspack(nextConfig)
 
 ## Usage with next-compose-plugins
 
-Alternatively, you can use `next-compose-plugins` to quickly integrate `@next/plugin-rspack` with other Next.js plugins:
+Alternatively, you can use `next-compose-plugins` to quickly integrate `next-rspack` with other Next.js plugins:
 
 ```js
 const withPlugins = require('next-compose-plugins')
-const withRspack = require('@next/plugin-rspack')
+const withRspack = require('next-rspack')
 
 module.exports = withPlugins([
   [withRspack],
