+++ packages/next/errors.json
@@ -645,8 +645,8 @@
   "644": "@rspack/core is not available. Please make sure the appropriate Next.js plugin is installed.",
   "645": "@rspack/plugin-react-refresh is not available. Please make sure the appropriate Next.js plugin is installed.",
   "646": "No span found for compilation",
-  "647": "@rspack/core is not available. Please make sure `@next/plugin-rspack` is correctly installed.",
-  "648": "@rspack/plugin-react-refresh is not available. Please make sure `@next/plugin-rspack` is correctly installed.",
+  "647": "@rspack/core is not available. Please make sure `next-rspack` is correctly installed.",
+  "648": "@rspack/plugin-react-refresh is not available. Please make sure `next-rspack` is correctly installed.",
   "649": "Cache handlers not initialized",
   "650": "experimental.nodeMiddleware",
   "651": "Unexpected module type %s",
