+++ packages/next/src/shared/lib/get-rspack.ts
@@ -8,7 +8,7 @@ export function getRspackCore() {
   } catch (e) {
     if (e instanceof Error && 'code' in e && e.code === 'MODULE_NOT_FOUND') {
       throw new Error(
-        '@rspack/core is not available. Please make sure `@next/plugin-rspack` is correctly installed.'
+        '@rspack/core is not available. Please make sure `next-rspack` is correctly installed.'
       )
     }
 
@@ -24,7 +24,7 @@ export function getRspackReactRefresh() {
   } catch (e) {
     if (e instanceof Error && 'code' in e && e.code === 'MODULE_NOT_FOUND') {
       throw new Error(
-        '@rspack/plugin-react-refresh is not available. Please make sure `@next/plugin-rspack` is correctly installed.'
+        '@rspack/plugin-react-refresh is not available. Please make sure `next-rspack` is correctly installed.'
       )
     }
